Ajuste metodo DataFactory e AplRelatorio.
Neste commit o método de gerar entrevistaRealizadaRandom e entrevistaRecusadaRandom
não estavam setando os atributos entrevistarealizada e doacaoautorizada como true.
Este problema foi corrigido para que os testes fossem mais precisos.

Na AplRelatorio foi acertado o metodo percentualEfeticacao para que esteve
retornasse um numero com 2 casas decimais de precisao. Para isso,o retorno
do método teve que ser alterado de Double para BigDecimal.

+++ src/main/java/br/ifes/leds/sincap/gerenciaNotificacao/cln/cdp/relatorios/TotalDoacaoInstituicao.java
@@ -3,6 +3,8 @@ package br.ifes.leds.sincap.gerenciaNotificacao.cln.cdp.relatorios;
 import lombok.*;
 import lombok.experimental.Builder;
 
+import java.math.BigDecimal;
+
 /**
  * @author aleao on 21/10/14.
  */
@@ -24,6 +26,6 @@ public class TotalDoacaoInstituicao {
 
     private Integer numeroDoacao;
 
-    private Double percentualEfetivacao;
+    private BigDecimal percentualEfetivacao;
 
 }
